Move controller+leaderelection into lib
