+++ lib/controller/controller_test.go
@@ -25,7 +25,7 @@ import (
 	"testing"
 	"time"
 
-	rl "github.com/kubernetes-incubator/nfs-provisioner/controller/leaderelection/resourcelock"
+	rl "github.com/kubernetes-incubator/nfs-provisioner/lib/leaderelection/resourcelock"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
 	fakev1core "k8s.io/client-go/kubernetes/typed/core/v1/fake"
