+++ lib/leaderelection/leaderelection.go
@@ -28,7 +28,7 @@ import (
 	"reflect"
 	"time"
 
-	rl "github.com/kubernetes-incubator/nfs-provisioner/controller/leaderelection/resourcelock"
+	rl "github.com/kubernetes-incubator/nfs-provisioner/lib/leaderelection/resourcelock"
 	"k8s.io/client-go/pkg/api/errors"
 	"k8s.io/client-go/pkg/api/unversioned"
 	"k8s.io/client-go/pkg/util/runtime"
