+++ lib/leaderelection/resourcelock/interface.go
